+++ src/test/java/com/deliveryfood/controller/CartControllerTest.java
@@ -1,9 +1,7 @@
 package com.deliveryfood.controller;
 
 import com.deliveryfood.common.mock.auth.WithAuthMember;
-import com.deliveryfood.dao.CartDao;
-import com.deliveryfood.model.request.CartMenuRequest;
-import com.deliveryfood.service.UserService;
+import com.deliveryfood.controller.model.request.CartMenuRequest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
