+++ src/test/java/com/deliveryfood/controller/UserControllerTest.java
@@ -1,11 +1,10 @@
 package com.deliveryfood.controller;
 
 import com.deliveryfood.common.mock.auth.WithAuthMember;
-import com.deliveryfood.model.request.UserRegisterRequest;
-import com.deliveryfood.model.request.UserRequest;
-import com.deliveryfood.model.request.UserUpdateRequest;
-import com.deliveryfood.service.MemberService;
-import com.deliveryfood.service.UserService;
+import com.deliveryfood.controller.model.request.UserRegisterRequest;
+import com.deliveryfood.controller.model.request.UserRequest;
+import com.deliveryfood.controller.model.request.UserUpdateRequest;
+import com.deliveryfood.service.impl.MemberService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
